Push utilities with lifecycle predicates to the package limit
They used to be right in the interface `EntityWithLifecycle` interface,
which repeated what the interface already said.

Also:
 * Reduced visibility of `AggregateEndpoint.dispatchInTx()` to package-private.
 * Simplified Javadocs.
