+++ server/src/main/java/io/spine/server/aggregate/AggregateTransaction.java
@@ -35,7 +35,6 @@ import io.spine.validate.ValidatingBuilder;
  * @param <I> the type of aggregate IDs
  * @param <S> the type of aggregate state
  * @param <B> the type of a {@code ValidatingBuilder} for the aggregate state
- * @author Alex Tymchenko
  */
 @Internal
 public class AggregateTransaction<I,
